--- src/features/chat/member/api/index.ts
@@ -1,5 +1,0 @@
-// API 함수들을 한 곳에서 export
-export { fetchChatMembers } from './members';
-export { createOneToOneRoom, fetchUserProfile, followUser, unfollowUser } from './profile';
-export { blockUser, checkIsGroupChat, reportUser } from './report';
-export { leaveChatRoom } from './room';
