--- src/features/chat/member/components/index.ts
@@ -1,8 +1,0 @@
-export { BottomSheetBase } from './BottomSheetBase';
-export { ChatMembersHeader } from './ChatMembersHeader';
-export { LeaveChatButton } from './LeaveChatButton';
-export { MemberItem } from './MemberItem';
-export { MembersList } from './MembersList';
-export { ReportBlockModal } from './ReportBlockModal';
-export { ReportDetailModal } from './ReportDetailModal';
-export { ReportReasonModal } from './ReportReasonModal';
