--- src/features/chat/room/api/userProfile.ts
@@ -1,2 +1,0 @@
-// shared로 이동됨 - 하위 호환성을 위해 re-export
-export { createOneToOneRoom, fetchUserProfile, followUser, unfollowUser } from '../../shared/api';
