--- src/features/chat/room/hooks/useUserProfile.ts
@@ -1,3 +1,0 @@
-// shared로 이동됨 - 하위 호환성을 위해 re-export
-export { useUserProfile } from '../../shared/hooks';
-export type { UserProfileActions, UserProfileHook, UserProfileState } from '../../shared/hooks';
