--- src/features/chat/room/hooks/useUserProfileQuery.ts
@@ -1,8 +1,0 @@
-// shared로 이동됨 - 하위 호환성을 위해 re-export
-export {
-  useCreateOneToOneRoomMutation,
-  useFollowUserMutation,
-  USER_PROFILE_QK,
-  useUnfollowUserMutation,
-  useUserProfileQuery,
-} from '../../shared/hooks';
