--- src/features/chat/shared/api/index.ts
@@ -1,1 +1,0 @@
-export { createOneToOneRoom, fetchUserProfile, followUser, unfollowUser } from './userProfile';
