--- src/features/chat/shared/hooks/index.ts
@@ -1,9 +1,0 @@
-export { useUserProfile } from './useUserProfile';
-export type { UserProfileActions, UserProfileHook, UserProfileState } from './useUserProfile';
-export {
-  useCreateOneToOneRoomMutation,
-  useFollowUserMutation,
-  USER_PROFILE_QK,
-  useUnfollowUserMutation,
-  useUserProfileQuery,
-} from './useUserProfileQuery';
