+++ src/features/chat/shared/hooks/useUserProfileQuery.ts
@@ -1,6 +1,6 @@
 import { CHAT_ROOMS_QUERY_KEY } from '@/src/features/chat/list/hooks/useChatRooms';
 import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
-import { createOneToOneRoom, fetchUserProfile, followUser, unfollowUser } from '../api';
+import { createOneToOneRoom, fetchUserProfile, followUser, unfollowUser } from '../api/userProfile';
 
 /** 쿼리 키 */
 export const USER_PROFILE_QK = (userId: number) => ['chat', 'userProfile', userId] as const;
